Start front-end favorite logic
Data seems to be coming in from the back-end appropriately.
Mapping to something similar to Course Cards will be necessary.

In addition, Context will need to be established to store Favorite
and some way of handling direct navigation (e.g., not from the
favorite button). Consider navigating to a sub-path, e.g.,
/favorites/my - inelegant but plausible.