Successfully add / view favorites without navigation control
This commit allows users to add to their favorites. Once confirmed
it will display a list of all of their favorites. However, we still
need to figure out how we will 1) store favorites and 3) how we will
handle direct navigation to /favorites, which at this point,
navigates to the confirmation screen.

consider changing the path to /favorites/add or navigationg to
favorites/my (more clumsy but perhaps easier to implement)